data type changed to double

+++ Lab-11/KenaKata/src/main/java/kenakata/catalog/Weighable.java
@@ -2,5 +2,5 @@ package kenakata.catalog;
 
 public interface Weighable
 {
-    long weightGrams();
+    double weightGrams();
 }
